+++ resolwe/flow/models/collection.py
@@ -1,6 +1,11 @@
 """Resolwe collection model."""
+import datetime
+
 from django.contrib.postgres.fields import ArrayField, JSONField
-from django.db import models
+from django.db import models, transaction
+
+from resolwe.permissions.shortcuts import get_objects_for_user
+from resolwe.permissions.utils import assign_contributor_permissions
 
 from .base import BaseModel
 from .utils import DirtyError, validate_schema
@@ -47,6 +52,18 @@ class BaseCollection(BaseModel):
         super().save()
 
 
+class CollectionQuerySet(models.QuerySet):
+    """Query set for ``Collection`` objects."""
+
+    @transaction.atomic
+    def duplicate(self, contributor=None):
+        """Duplicate (make a copy) ``Collection`` objects."""
+        return [
+            collection.duplicate(contributor=contributor)
+            for collection in self
+        ]
+
+
 class Collection(BaseCollection):
     """Postgres model for storing a collection."""
 
@@ -61,3 +78,42 @@ class Collection(BaseCollection):
             ("add_collection", "Can add data objects to collection"),
             ("owner_collection", "Is owner of the collection"),
         )
+
+    #: manager
+    objects = CollectionQuerySet.as_manager()
+
+    #: duplication date and time
+    duplicated = models.DateTimeField(blank=True, null=True)
+
+    def is_duplicate(self):
+        """Return True if collection is a duplicate."""
+        return bool(self.duplicated)
+
+    def duplicate(self, contributor=None):
+        """Duplicate (make a copy)."""
+        duplicate = Collection.objects.get(id=self.id)
+        duplicate.pk = None
+        duplicate.slug = None
+        duplicate.name = 'Copy of {}'.format(self.name)
+        duplicate.duplicated = datetime.datetime.now()
+        if contributor:
+            duplicate.contributor = contributor
+
+        duplicate.save(force_insert=True)  # pylint: disable=protected-access
+
+        assign_contributor_permissions(duplicate)
+
+        # Fields to inherit from original data object.
+        duplicate.created = self.created
+        duplicate.save()
+
+        # Duplicate collection's entities.
+        entities = get_objects_for_user(contributor, 'view_entity', self.entity_set.all())  # pylint: disable=no-member
+        duplicated_entities = entities.duplicate(contributor=contributor)
+        duplicate.entity_set.add(*duplicated_entities)
+
+        # Add duplicated data objects to collection.
+        for duplicated_entity in duplicate.entity_set.all():  # pylint: disable=no-member
+            duplicate.data.add(*duplicated_entity.data.all())
+
+        return duplicate