+++ docs/CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
   of a new process from the running Python process
 - Add DataLocation model and pair it with Data model to handle data location
 - Add ``entity_names`` field to ``DataSerializer``
+- Support duplication of ``Data``, ``Entity`` and ``Collection``
 
 
 ===================